minor refactor; delete unused classes and methods; mark UserRepository and UserService as Deprecated
