+++ src/main/java/com/andersentask/bookshop/http/servlets/ChangeBookStatusServlet.java
@@ -1,7 +1,7 @@
 package com.andersentask.bookshop.http.servlets;
 
 import com.andersentask.bookshop.book.enums.BookStatus;
-import com.andersentask.bookshop.console.Commands;
+import com.andersentask.bookshop.broker.Commands;
 
 import java.util.List;
 import java.util.Locale;
