+++ src/main/java/com/andersentask/bookshop/http/servlets/CreateRequestServlet.java
@@ -1,6 +1,6 @@
 package com.andersentask.bookshop.http.servlets;
 
-import com.andersentask.bookshop.console.Commands;
+import com.andersentask.bookshop.broker.Commands;
 
 import java.util.Map;
 
@@ -16,6 +16,6 @@ public class CreateRequestServlet extends JsonServlet {
     Response get(String uri, Map<String, String[]> parameters) {
         return new Response
                 (commands.createRequest
-                        (Long.parseLong(parameters.getOrDefault("bookId",new String[]{""})[0])));
+                        (Long.parseLong(parameters.getOrDefault("bookId", new String[]{""})[0])));
     }
 }
