+++ src/main/java/com/andersentask/bookshop/order/enums/OrderStatus.java
@@ -3,5 +3,5 @@ package com.andersentask.bookshop.order.enums;
 public enum OrderStatus {
     IN_PROCESS,
     COMPLETED,
-    CANCELED;
+    CANCELED
 }
