+++ src/main/java/com/andersentask/bookshop/order/repositories/OrderMapper.java
@@ -38,5 +38,4 @@ class OrderMapper {
 
     private OrderMapper() {
     }
-
 }
