+++ src/main/java/com/andersentask/bookshop/order/repositories/OrderRepository.java
@@ -9,7 +9,12 @@ import lombok.extern.slf4j.Slf4j;
 
 import javax.sql.DataSource;
 import java.math.BigDecimal;
-import java.sql.*;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Statement;
+import java.sql.Timestamp;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -225,5 +230,4 @@ public class OrderRepository {
             throw new IllegalArgumentException(e);
         }
     }
-
 }
