+++ src/main/java/com/andersentask/bookshop/request/repository/RequestMapper.java
@@ -22,6 +22,6 @@ class RequestMapper {
         }
     };
 
-    private RequestMapper(){
+    private RequestMapper() {
     }
 }
