+++ src/main/java/com/andersentask/bookshop/request/repository/RequestRepository.java
@@ -6,7 +6,11 @@ import com.andersentask.bookshop.request.entities.Request;
 import lombok.extern.slf4j.Slf4j;
 
 import javax.sql.DataSource;
-import java.sql.*;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
